CHE-5473: move implementation of JUnit4 to the special package
